(PC-13482)[ADAGE] tech: move getLocalDatetime from OfferSummary to date module
